requirements.txt
fastapi
uvicorn
requests
aiogram
python-dotenv
